utilisateur

+++ vehicule/views.py
@@ -90,12 +90,14 @@ def modifier_vehicule(request, pk):
     if request.method == 'POST':
         form = VehiculeForm(request.POST, request.FILES, instance=vehicule)
         if form.is_valid():
-            # Supprimez les anciennes images du véhicule
-            Photo.objects.filter(vehicule=vehicule).delete()
+            if request.FILES.getlist('images'):
+                # Supprimez les anciennes images du véhicule
+                Photo.objects.filter(vehicule=vehicule).delete()
 
-            # Ajoutez les nouvelles images au véhicule
-            for image in request.FILES.getlist('images'):
-                Photo.objects.create(vehicule=vehicule, images=image)
+            if request.FILES.getlist('images'):
+                # Ajoutez les nouvelles images au véhicule
+                for image in request.FILES.getlist('images'):
+                    Photo.objects.create(vehicule=vehicule, images=image)
 
             # Enregistrez le formulaire du véhicule mis à jour
             form.save()
